U-106, Added the table for professors; fixed details in schedules and separated the classes Professor and Schedule into different files
