+++ src/app/models/period.model.ts
@@ -1,3 +1,6 @@
+import {Schedule} from './schedule.model';
+import {Professor} from './professor.model';
+
 export class Period {
   public schedules: Schedule[];
   public professors: Professor[];
@@ -15,25 +18,3 @@ export class Period {
   }
 }
 
-export class Schedule {
-  public dayOfWeek: string;
-  public beginTime: number;
-  public endTime: number;
-  public classroom: string;
-
-  constructor(dayOfWeek?: string,
-              beginTime?: number,
-              endTime?: number,
-              classroom?: string) {
-    this.dayOfWeek = dayOfWeek;
-    this.beginTime = beginTime;
-    this.endTime = endTime;
-    this.classroom = classroom;
-  }
-}
-
-
-export class Professor {
-  public name: string;
-  public lastName: string;
-}
