+++ src/app/models/professor.model.ts
@@ -1,0 +1,9 @@
+export class Professor {
+  public name: string;
+  public lastName: string;
+
+  constructor(name: string, lastName: string) {
+    this.name = name;
+    this.lastName = lastName;
+  }
+}
