+++ src/app/modules/course/course.module.ts
@@ -5,12 +5,14 @@ import {CourseComponent} from './pages/course/course.component';
 import {SharedModule} from '../../shared/shared.module';
 import { PeriodComponent } from './components/period/period.component';
 import { SchedulesComponent } from './components/schedules/schedules.component';
+import { ProfessorsComponent } from './components/teachers/professors.component';
 
 @NgModule({
   declarations: [
     CourseComponent,
     PeriodComponent,
-    SchedulesComponent
+    SchedulesComponent,
+    ProfessorsComponent
   ],
   imports: [
     CommonModule,
