+++ src/styles.css
@@ -31,3 +31,6 @@
 form.submitted input.ng-invalid, input.ng-dirty.ng-invalid {
   border: 1px solid red;
 }
+.margin-top-2{
+  margin-top: 2rem;
+}
